requirements.txt
colorama==0.4.3
mysql-connector-python==8.0.21
protobuf==3.13.0
six==1.15.0
